♻️ ooconnector: routing refactor (#525)
